requirements.txt
stellar_sdk
alive_progress
colorama
python-dotenv